feat: add exception throw

+++ Infrastructure/Persistence/Repositories/ClientRepository.cs
@@ -45,14 +45,31 @@ public class ClientRepository : IClientRepository
 
     public async Task DeleteClientById(Guid id)
     {
+        bool pending = await _context.Payments
+            .AnyAsync(p => p.SenderId == id && p.Status == Domain.Enums.PaymentStatus.PENDING);
+
+        if (pending) 
+            throw new BusinessException("CLient have a payment pending");
 
         await using var transaction = await _context.Database.BeginTransactionAsync();
 
-        var rowsAffected = await _context.Clients
+        try
+        {
+            var rowsAffected = await _context.Clients
                               .Where(c => c.Id == id)
                               .ExecuteDeleteAsync();
 
-        if (rowsAffected == 0) throw new ClientNotFoundException();
+            if (rowsAffected == 0) throw new ClientNotFoundException();
+
+            if (rowsAffected > 1) throw new UnexpectedException("Multiple users affected");
+
+            await transaction.CommitAsync();
+        }
+        catch
+        {
+            throw;
+        }
+        
 
     }
 }
